trace/: move connection-keeping code to its own file
